requirements.txt
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.24.0
matplotlib>=3.7.0
pandas>=2.0.0
pillow>=10.0.0
streamlit>=1.28.0
streamlit-drawable-canvas>=0.9.0
scikit-learn>=1.3.0
seaborn>=0.12.0
pytest>=7.4.0
black>=23.0.0
flake8>=6.0.0